Attempt to improve performance.
It's obvious that cat() for each write is really sub optimal (opens/closes
file each time).

Unfortunately, on Linux at least (not tested on Windows) constructing the
entire file as a string with paste() is extremely slow.  But calling cat()
each time was pretty snappy.  Unfortunately, on Windows calling cat() each
time is painfully slow.

So as a compromise we now buffer each line of the image before writing.

This is very snappy on Windows and presumably also on Linux.